Extract Undefined to a separate file
