+++ concrete_settings/__init__.py
@@ -1,5 +1,4 @@
 import sys
-
 from .concrete_settings import (
     Settings,
     Setting,
