+++ concrete_settings/undefined.py
@@ -1,0 +1,14 @@
+class UndefinedMeta(type):
+    def __bool__(self):
+        return False
+
+    def __str__(self):
+        return 'Undefined value'
+
+
+class Undefined(metaclass=UndefinedMeta):
+    """`Undefined` is a special value which indicates
+    that something has not been explicitly set by a user.
+    """
+
+    pass
